[wip] Ensure that tab/WebContents objects are added to memory as soon as they are created.
In the case of tabs which get discarded, they are replaced with a fresh, clean WebContents. We were not receiving this new object, so when the renderer asked for this new object's tabId to be made active, the cache did not have that object.
We must also handle updating references to the old TabId to the new TabId, when remembering opener tab IDs and historical tab 'active' history for a window.

TODO: discarded tab never gets added to tabState
TODO: fix moving a discarded tab to a new window (the detached event never fires after being detached, the 'new' tab gets destroyed immediately, and the original discarded tab is not in tabState)