Avoid blank tabs which never have guest-ready set in an in-transit frame
guest-ready could be sent to the original window whilst the frame object has already been sent to the browser / new window without the `guestIsReady` property.
Instead of moving this to the browser process, I've removed the event handling entirely as it does not seem like we need to wait for guest-ready anymore (we weren't for attaching the Tab for preview mode anyway...).